write compressed data to file, read it and decompress

+++ huffman/node.go
@@ -3,7 +3,7 @@ package huffman
 //Node ...
 type Node struct {
 	Charachter string
-	Weight     uint64
+	Weight     uint
 	CodePath   *BitArray
 	Nodes      []*Node
 }
